# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CertificateListResponse:
    properties:
      certificates:
        items:
          $ref: '#/definitions/dto.CertificateResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.CertificateRequest:
    properties:
      branch_id:
        type: string
      certificate_data:
        items:
          type: integer
        type: array
      certificate_path:
        type: string
      expiration_date:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
    required:
    - branch_id
    - expiration_date
    - name
    - password
    type: object
  dto.CertificateResponse:
    properties:
      branch_id:
        type: string
      branch_name:
        type: string
      certificate_path:
        type: string
      created_at:
        type: string
      expiration_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_expired:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      message:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8084
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para o sistema ERP de supermercados
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SuperERP API
  version: "1.0"
paths:
...
  /certificates:
    get:
      description: Lista os certificados com paginação
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (padrão: 10)'
        in: query
        name: page_size
        type: integer
      - description: Filtrar por filial
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CertificateListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Listar certificados
      tags:
      - Certificados
    post:
      consumes:
      - application/json
      description: Cria um novo certificado digital
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados do certificado
        in: body
        name: certificate
        required: true
        schema:
          $ref: '#/definitions/dto.CertificateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CertificateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Criar certificado
      tags:
      - Certificados
  /certificates/{id}:
    delete:
      description: Remove um certificado pelo ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID do certificado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Excluir certificado
      tags:
      - Certificados
    get:
      description: Busca um certificado pelo ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID do certificado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CertificateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obter certificado
      tags:
      - Certificados
    put:
      consumes:
      - application/json
      description: Atualiza um certificado existente
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID do certificado
        in: path
        name: id
        required: true
        type: string
      - description: Dados do certificado
        in: body
        name: certificate
        required: true
        schema:
          $ref: '#/definitions/dto.CertificateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CertificateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Atualizar certificado
      tags:
      - Certificados
  /certificates/{id}/activate:
    post:
      description: Ativa um certificado pelo ID (e desativa outros certificados da
        mesma filial)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID do certificado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Ativar certificado
      tags:
      - Certificados
  /certificates/{id}/deactivate:
    post:
      description: Desativa um certificado pelo ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID do certificado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Desativar certificado
      tags:
      - Certificados
  /certificates/expiring:
    get:
      description: Lista os certificados que expirarão em X dias
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Número de dias (padrão: 30)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CertificateListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Listar certificados expirando
      tags:
      - Certificados
  /certificates/upload:
    post:
      consumes:
      - multipart/form-data
      description: Realiza upload de um arquivo de certificado digital
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Arquivo do certificado (.pfx)
        in: formData
        name: file
        required: true
        type: file
      - description: ID da filial
        in: formData
        name: branch_id
        required: true
        type: string
      - description: Nome do certificado
        in: formData
        name: name
        required: true
        type: string
      - description: Senha do certificado
        in: formData
        name: password
        required: true
        type: string
      - description: Data de validade (YYYY-MM-DD)
        in: formData
        name: expiration_date
        required: true
        type: string
      - description: Se o certificado está ativo
        in: formData
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CertificateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Upload de certificado
      tags:
      - Certificados
...
securityDefinitions:
  Bearer:
    description: 'Cabeçalho de autenticação JWT usando o esquema Bearer. Exemplo:
      "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"